+++ lib/node.ts
@@ -1,12 +1,6 @@
 /* eslint-disable @typescript-eslint/no-base-to-string */
 import {cache} from '../util/lint';
-import {
-	Shadow,
-
-	/* NOT FOR BROWSER */
-
-	typeError,
-} from '../util/debug';
+import {Shadow} from '../util/debug';
 import type {LintError, AstNode as AstNodeBase, TokenTypes} from '../base';
 import type {Cached} from '../util/lint';
 import type {
@@ -474,7 +468,11 @@ export abstract class AstNode implements AstNodeBase {
 	/* istanbul ignore next */
 	/** @private */
 	typeError(method: string, ...types: string[]): never {
-		return typeError(this.constructor, method, ...types);
+		throw new TypeError(
+			`${this.constructor.name}.${method} method only accepts ${
+				types.join(', ')
+			} as input parameters!`,
+		);
 	}
 
 	/** @private */